requirements.txt
Flask>=2.0
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
Flask-JWT-Extended
passlib[bcrypt]
